Refactoring script.js into mulitple files
