+++ static/js/uiElements.js
@@ -1,0 +1,3 @@
+export const chatMessages = document.getElementById('chat-messages');
+export const userInput = document.getElementById('user-input');
+export const sendButton = document.getElementById('send-button');
