+++ templates/index.html
@@ -20,6 +20,6 @@
         </div>
     </div>
 
-    <script src="{% static 'js/script.js' %}"></script>
+    <script type="module" src="{% static 'js/main.js' %}"></script>
 </body>
 </html>
